requirements.txt
pytest==7.4.3
selene==2.0.0rc4
allure-pytest==2.9.45
python-dotenv==0.20.0
requests == 2.31.0
jsonschema
pydantic-settings